deploy/requirements.txt
tensorflow-cpu
numpy~=1.19.2
pillow
gradio
wget